+++ messages/responses.go
@@ -21,6 +21,7 @@ var responses = map[byte]func() Response{
 	0x82: func() Response { return new(GetDoorControlStateResponse) },
 	0x88: func() Response { return new(SetTimeProfileResponse) },
 	0x8a: func() Response { return new(ClearTimeProfilesResponse) },
+	0x8c: func() Response { return new(SetSuperPasswordResponse) },
 	0x8e: func() Response { return new(RecordSpecialEventsResponse) },
 	0x90: func() Response { return new(SetListenerResponse) },
 	0x92: func() Response { return new(GetListenerResponse) },