+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+1. Implemented `set-super-password` function.
+
 
 ## [0.8.6](https://github.com/uhppoted/uhppote-core/releases/tag/v0.8.6) - 2023-08-30